Refactor InterfaceRuleType to extend MultiRuleType and update calculate method to handle multiple sources

+++ src/main/java/es/vodafone/sid/poller/model/Source.java
@@ -27,7 +27,7 @@ public record Source(
     String capture,
     Instant instant,
     BigInteger cache,
-    Integer scale,
+    Short scale,
     Boolean active
 ) {
   public boolean isMulti() {
